SBERDOMA-992 Move loadModelsByChunks to common server utils fix export template
